chore(fee, concurrency): move add_fee_to_sequencer_balance to fee_utils
