+++ crates/blockifier/src/concurrency/fee_utils.rs
@@ -1,6 +1,18 @@
-use starknet_api::hash::StarkFelt;
+use std::collections::HashMap;
 
+use num_traits::ToPrimitive;
+use starknet_api::core::ContractAddress;
+use starknet_api::hash::StarkFelt;
+use starknet_api::stark_felt;
+use starknet_api::transaction::Fee;
+
+use crate::context::BlockContext;
 use crate::execution::call_info::CallInfo;
+use crate::execution::execution_utils::stark_felt_to_felt;
+use crate::fee::fee_utils::get_sequencer_balance_keys;
+use crate::state::cached_state::{ContractClassMapping, StateMaps};
+use crate::state::state_api::UpdatableState;
+
 #[cfg(test)]
 #[path = "fee_utils_test.rs"]
 mod test;
@@ -27,4 +39,36 @@ pub fn fill_sequencer_balance_reads(
     }
     storage_read_values[low_index] = sequencer_balance_low;
     storage_read_values[high_index] = sequencer_balance_high;
+}
+
+pub fn add_fee_to_sequencer_balance(
+    fee_token_address: ContractAddress,
+    state: &mut impl UpdatableState,
+    actual_fee: Fee,
+    block_context: &BlockContext,
+    sequencer_balance_value_low: StarkFelt,
+    sequencer_balance_value_high: StarkFelt,
+) {
+    let sequencer_balance_low_as_u128 = stark_felt_to_felt(sequencer_balance_value_low)
+        .to_u128()
+        .expect("sequencer balance low should be u128");
+    let sequencer_balance_high_as_u128 = stark_felt_to_felt(sequencer_balance_value_high)
+        .to_u128()
+        .expect("sequencer balance high should be u128");
+    let (new_value_low, carry) = sequencer_balance_low_as_u128.overflowing_add(actual_fee.0);
+    let (new_value_high, carry) = sequencer_balance_high_as_u128.overflowing_add(carry.into());
+    assert!(
+        !carry,
+        "The sequencer balance overflowed when adding the fee. This should not happen."
+    );
+    let (sequencer_balance_key_low, sequencer_balance_key_high) =
+        get_sequencer_balance_keys(block_context);
+    let writes = StateMaps {
+        storage: HashMap::from([
+            ((fee_token_address, sequencer_balance_key_low), stark_felt!(new_value_low)),
+            ((fee_token_address, sequencer_balance_key_high), stark_felt!(new_value_high)),
+        ]),
+        ..StateMaps::default()
+    };
+    state.apply_writes(&writes, &ContractClassMapping::default(), &HashMap::default());
 }
