+++ crates/blockifier/src/concurrency/worker_logic.rs
@@ -4,23 +4,18 @@ use std::sync::Mutex;
 use std::thread;
 use std::time::Duration;
 
-use num_traits::ToPrimitive;
-use starknet_api::core::{ClassHash, ContractAddress};
-use starknet_api::hash::StarkFelt;
-use starknet_api::stark_felt;
+use starknet_api::core::ClassHash;
 use starknet_api::transaction::Fee;
 
-use super::versioned_state::{VersionedState, VersionedStateProxy};
+use super::versioned_state::VersionedState;
 use crate::blockifier::transaction_executor::TransactionExecutorError;
 use crate::bouncer::Bouncer;
-use crate::concurrency::fee_utils::fill_sequencer_balance_reads;
+use crate::concurrency::fee_utils::{add_fee_to_sequencer_balance, fill_sequencer_balance_reads};
 use crate::concurrency::scheduler::{Scheduler, Task};
 use crate::concurrency::utils::lock_mutex_in_array;
 use crate::concurrency::versioned_state::ThreadSafeVersionedState;
 use crate::concurrency::TxIndex;
 use crate::context::BlockContext;
-use crate::execution::execution_utils::stark_felt_to_felt;
-use crate::fee::fee_utils::get_sequencer_balance_keys;
 use crate::state::cached_state::{
     ContractClassMapping, StateChanges, StateMaps, TransactionalState,
 };
@@ -312,38 +307,4 @@ impl<'a, U: UpdatableState> WorkerExecutor<'a, U> {
             .into_inner_state()
             .commit_chunk_and_recover_block_state(n_committed_txs, visited_pcs)
     }
-}
-
-// Utilities.
-
-fn add_fee_to_sequencer_balance(
-    fee_token_address: ContractAddress,
-    tx_versioned_state: &mut VersionedStateProxy<impl StateReader>,
-    actual_fee: Fee,
-    block_context: &BlockContext,
-    sequencer_balance_value_low: StarkFelt,
-    sequencer_balance_value_high: StarkFelt,
-) {
-    let sequencer_balance_low_as_u128 = stark_felt_to_felt(sequencer_balance_value_low)
-        .to_u128()
-        .expect("sequencer balance low should be u128");
-    let sequencer_balance_high_as_u128 = stark_felt_to_felt(sequencer_balance_value_high)
-        .to_u128()
-        .expect("sequencer balance high should be u128");
-    let (new_value_low, carry) = sequencer_balance_low_as_u128.overflowing_add(actual_fee.0);
-    let (new_value_high, carry) = sequencer_balance_high_as_u128.overflowing_add(carry.into());
-    assert!(
-        !carry,
-        "The sequencer balance overflowed when adding the fee. This should not happen."
-    );
-    let (sequencer_balance_key_low, sequencer_balance_key_high) =
-        get_sequencer_balance_keys(block_context);
-    let writes = StateMaps {
-        storage: HashMap::from([
-            ((fee_token_address, sequencer_balance_key_low), stark_felt!(new_value_low)),
-            ((fee_token_address, sequencer_balance_key_high), stark_felt!(new_value_high)),
-        ]),
-        ..StateMaps::default()
-    };
-    tx_versioned_state.apply_writes(&writes, &ContractClassMapping::default(), &HashMap::default());
 }
